+++ sdk/cosmos/azure-cosmos-test/src/main/java/com/azure/cosmos/test/implementation/faultinjection/RntbdServerErrorInjector.java
@@ -76,7 +76,7 @@ public class RntbdServerErrorInjector implements IRntbdServerErrorInjector {
         RxDocumentServiceRequest request = requestRecord.args().serviceRequest();
         FaultInjectionServerErrorRule serverConnectionDelayRule = this.ruleStore.findRntbdServerConnectionDelayRule(request);
 
-        if(serverConnectionDelayRule != null) {
+        if (serverConnectionDelayRule != null) {
             request.faultInjectionRequestContext
                 .applyFaultInjectionRule(
                     requestRecord.transportRequestId(),