+++ sdk/cosmos/azure-cosmos-test/CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 1.0.0-beta.1 (Unreleased)
 
 #### Features Added
-* Added fault injection support. 
+* Added fault injection support - See [PR 33329](https://github.com/Azure/azure-sdk-for-java/pull/33329) 
 
 #### Breaking Changes